+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "agnai",
-  "version": "0.1.0-alpha-21",
+  "version": "0.1.0-alpha-22",
   "description": "Agnostic AI Chat",
   "keywords": [],
   "license": "AGPL-3.0",
