+++ web/store/data/image.ts
@@ -1,7 +1,9 @@
 import * as horde from '../../../common/horde-gen'
 import { createImagePrompt } from '../../../common/image-prompt'
 import { api, isLoggedIn } from '../api'
-import { getPromptEntities } from './messages'
+import { subscribe } from '../socket'
+import { PromptEntities, getPromptEntities, msgsApi } from './messages'
+import { AIAdapter } from '/common/adapters'
 
 type GenerateOpts = {
   chatId?: string
@@ -17,7 +19,7 @@ export const imageApi = {
 
 export async function generateImage({ chatId, messageId, onDone, ...opts }: GenerateOpts) {
   const entities = await getPromptEntities()
-  const prompt = opts.prompt ? opts.prompt : await createImagePrompt(entities)
+  const prompt = opts.prompt ? opts.prompt : await createSummarizedImagePrompt(entities)
 
   if (!isLoggedIn()) {
     const image = await horde.generateImage(entities.user, prompt)
@@ -31,4 +33,34 @@ export async function generateImage({ chatId, messageId, onDone, ...opts }: Gene
     ephemeral: opts.ephemeral,
   })
   return res
+}
+
+const SUMMARY_BACKENDS: { [key in AIAdapter]?: boolean } = {
+  openai: true,
+}
+
+async function createSummarizedImagePrompt(opts: PromptEntities) {
+  if (opts.settings?.service! in SUMMARY_BACKENDS && opts.user.images?.summariseChat) {
+    console.log('Using', opts.settings?.service, 'to summarise')
+    msgsApi.generateResponseV2({ kind: 'summary' })
+
+    return new Promise<string>((resolve, reject) => {
+      let timer = setTimeout(() => {
+        reject(new Error(`Chat summarisation timed out`))
+      }, 45000)
+      subscribe(
+        'chat-summary',
+        { chatId: 'string', summary: 'string' },
+        (body) => {
+          console.log('received summary')
+          clearTimeout(timer)
+          resolve(body.summary)
+        },
+        true
+      )
+    })
+  }
+
+  const prompt = await createImagePrompt(opts)
+  return prompt
 }
